scripts/sentiment_analysis.py
import sqlite3

def create_sentiment_results_table(conn):
    # Create sentiment_results table if not exists
    c = conn.cursor()
    c.execute('''CREATE TABLE IF NOT EXISTS sentiment_results (
                id INTEGER PRIMARY KEY,
                review_id INTEGER,
                sentiment TEXT
                )''')
    conn.commit()

def analyze_sentiment(conn_reviews, conn_results):
    # Example function to analyze sentiment (implementation depends on your requirements)
    c_reviews = conn_reviews.cursor()
    c_results = conn_results.cursor()
    
    # Example query to fetch reviews
    c_reviews.execute('SELECT * FROM reviews')
    reviews = c_reviews.fetchall()
    
    # Example sentiment analysis and storing results
    for review in reviews:
        review_id = review[0]
        review_text = review[2]
        
        # Perform sentiment analysis (example: using TextBlob)
        sentiment = analyze_sentiment_textblob(review_text)
        
        # Insert sentiment result into sentiment_results table
        c_results.execute("INSERT INTO sentiment_results (review_id, sentiment) VALUES (?, ?)", (review_id, sentiment))
    
    conn_results.commit()

def analyze_sentiment_textblob(text):
    # Example sentiment analysis function using TextBlob (you can replace with your own implementation)
    from textblob import TextBlob
    blob = TextBlob(text)
    sentiment = blob.sentiment.polarity
    
    if sentiment > 0:
        return 'positive'
    elif sentiment < 0:
        return 'negative'
    else:
        return 'neutral'

# Example usage (you can remove this from the script if not needed)
if __name__ == '__main__':
    conn_reviews = sqlite3.connect('data/reviews.db')  # Adjust path as needed
    conn_results = sqlite3.connect('data/results.db')  # Adjust path as needed
    
    create_sentiment_results_table(conn_results)
    analyze_sentiment(conn_reviews, conn_results)
    
    conn_reviews.close()
    conn_results.close()
